Refactor server port calculation
This commit changes the way the server port calculation is performed.
In order of precidence, it will calculate a dynamic port if required,
otherwise it will obtain the server port from the command line args,
otherwise it will obtain the server port from the application
properties, and if none of those options are valid 8080 will be
returned.